+++ src/lib/views/DataView/actions/api.js
@@ -3,6 +3,18 @@ import { api, string } from '~su/utilities'
 
 import { getApiConfig } from '~su/store/root-store'
 
+const extractFiltersSchema = (json_schema) => {
+  if (!json_schema?.properties?.filters) {
+    return null
+  }
+
+  const { filters, order } = json_schema.properties
+
+  filters.properties.order = order
+
+  return filters
+}
+
 const setupApiActions = (
   { setState, setData, itemName, itemPluralName, collectionApiPath, apiPlaceholders = {} },
   messageApi
@@ -18,13 +30,13 @@ const setupApiActions = (
     const apiConfig = getApiConfig()
 
     // flush the store
-    setData({ [itemPluralName]: [], pagination: {}, filters: {} })
+    setData({ [itemPluralName]: [], pagination: {}, filtersSchema: { type: 'object', properties: {}, required: [] } })
 
     // construct path
     const path = string.replacePlaceholders(`${apiConfig[collectionApiPath]}${urlSearch}`, apiPlaceholders)
 
     const onSuccess = ({ json_schema }) => {
-      setState('filtersSchema', json_schema.properties.filters)
+      setState('filtersSchema', extractFiltersSchema(json_schema))
     }
 
     loadCollection({ path, onSuccess })